+++ nasnav-services/src/main/java/com/nasnav/service/LoyaltyCoinsDropServiceImp.java
@@ -14,6 +14,7 @@ import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.chrono.HijrahDate;
@@ -157,7 +158,7 @@ public class LoyaltyCoinsDropServiceImp implements LoyaltyCoinsDropService {
         LoyaltyCoinsDropEntity entity = loyaltyCoinsDropRepository.getByOrganization_IdAndTypeId(orgId, typeId);
         //Create Coins Drop Log
         loyaltyCoinsDropLogsService.updateCoinsDropLog(entity);
-        Integer amount = entity.getAmount();
+        BigDecimal amount = entity.getAmount();
 
         ShopsEntity shopEntity = shopsRepository.findById(shopId).get();
